pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-twin-ui"
version = "0.1.0"
description = "Digital Twin User Interface — simulation automation platform with DOE, MLflow, and ML prediction"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "lxml>=5.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "scikit-learn>=1.4.0",
    "torch>=2.3.0",
    "mlflow>=2.13.0",
    "scipy>=1.13.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "loguru>=0.7.2",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["digital_twin_ui*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that use the full simulation template and are slower",
]

[tool.coverage.run]
source = ["digital_twin_ui"]
omit = ["tests/*"]
